requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
cycler==0.11.0
docopt==0.6.2
fonttools==4.28.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.22.3
packaging==21.3
Pillow==8.4.0
pipreqs==0.4.11
pycodestyle==2.8.0
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
QtPy==2.0.1
requests==2.27.1
setuptools-scm==6.3.2
six==1.16.0
toml==0.10.2
tomli==1.2.2
urllib3==1.26.9
yarg==0.1.9
